Create a separate fly_app_secret resource
Instead of managing secrets as part of managing an app, create
a separate resource.
This allows creating secrets that are dependent on a resource that's in
turn dependent on the app the secrets are being defined it.
It's also inline with the expectation of one resource == one API call.

Manage a single resource per secret to simplify update logic, at the
cost of more API calls.